+++ packages/core/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [42.9.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.8.0...@wireapp/core@42.9.0) (2023-10-02)
+
+### Features
+
+* specify initial nb of key packages in corecrypto ([#5572](https://github.com/wireapp/wire-web-packages/issues/5572)) ([e603324](https://github.com/wireapp/wire-web-packages/commit/e603324a963673b4eec69ddb3d4b86d56fc5f1c6))
+
+# [42.8.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.7.0...@wireapp/core@42.8.0) (2023-09-28)
+
+### Features
+
+* recover from mls stale message error ([#5497](https://github.com/wireapp/wire-web-packages/issues/5497)) ([7d70999](https://github.com/wireapp/wire-web-packages/commit/7d7099968b9e92a93974da683400868ed04cd535))
+
+# [42.7.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.6.2...@wireapp/core@42.7.0) (2023-09-26)
+
+### Features
+
+* move TypedEventEmitter to commons package ([#5551](https://github.com/wireapp/wire-web-packages/issues/5551)) ([989bc47](https://github.com/wireapp/wire-web-packages/commit/989bc47e649cbc9c2addaf3504f654b8bf44620f))
+
+## [42.6.2](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.6.1...@wireapp/core@42.6.2) (2023-09-25)
+
+**Note:** Version bump only for package @wireapp/core
+
+## [42.6.1](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.6.0...@wireapp/core@42.6.1) (2023-09-25)
+
+**Note:** Version bump only for package @wireapp/core
+
 # [42.6.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/core@42.5.1...@wireapp/core@42.6.0) (2023-09-25)
 
 ### Features